Combine login and PhotonManager
