+++ Patriot/Hardware/LoggingIn.swift
