+++ Patriot/View Controllers/LoginViewController.swift
@@ -66,10 +66,15 @@ class LoginViewController: UIViewController {
             loginManager.login(user: user, password: password) { (error) in
                 if let error = error {
                     print("Error auto logging in: \(error)")
+                    //TODO: display an error message, but stay on this screen
+                    return
                 }
+                //TODO: start device discovery
+                print("Starting device discovery...")
+                
                 self.updateDisplay()
 
-                //TODO: Delay 1 second then switch to Activities display after auto logging in
+                //Delay 1 second then switch to Activities display after auto logging in
                 DispatchQueue.main.asyncAfter(deadline: .now() + 0.75) {
                     if let tabBarController = self.tabBarController as? TabBarController {
                         tabBarController.selectActivitiesTab()
